isvtec_oauth_provider/requirements.txt
Django==1.3
django-oauth-plus==2.0
hg+https://bitbucket.org/ubernostrum/django-registration
django-registration-defaults==0.3
django-social-auth==0.6.0
oauth2==1.5.170
python-openid==2.2.5
